[zutils] F: Allow relative imports
